refactor2
