+++ CPE200Lab1/CPE200Lab1/BasicCalculatorEngine.cs
@@ -1,0 +1,160 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CPE200Lab1
+{
+    public class BasicCalculatorEngine
+    {
+        public bool isNumber(string str)
+        {
+            double retNum;
+            return Double.TryParse(str, out retNum);
+        }
+        public bool isOperator(string str)
+        {
+            switch (str)
+            {
+                case "+":
+                case "-":
+                case "X":
+                case "÷":
+                case "%":
+                    return true;
+            }
+            return false;
+        }
+        public bool isUnary(string str)
+        {
+            switch (str)
+            {
+                case "√":
+                case "1/x":
+                    return true;
+            }
+            return false;
+        }
+
+        public string calculate(string oper, string firstOperand, int maxOutputSize = 8)
+        {
+            try
+            {
+                switch (oper)
+                {
+                    case "√":
+                        {
+                            double result;
+                            string[] parts;
+                            int remainLength;
+
+                            result = Math.Sqrt(Convert.ToDouble(firstOperand));
+                            // split between integer part and fractional part
+                            parts = result.ToString().Split('.');
+                            // if integer part length is already break max output, return error
+                            if (parts[0].Length > maxOutputSize)
+                            {
+                                return "E";
+                            }
+                            // calculate remaining space for fractional part.
+                            remainLength = maxOutputSize - parts[0].Length - 1;
+                            // trim the fractional part gracefully. 
+                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
+
+                        }
+                    case "1/x":
+                        if (firstOperand != "0")
+                        {
+                            double result;
+                            string[] parts;
+                            int remainLength;
+
+                            result = (1.0 / Convert.ToDouble(firstOperand));
+                            // split between integer part and fractional part
+                            parts = result.ToString().Split('.');
+                            // if integer part length is already break max output, return error
+                            if (parts[0].Length > maxOutputSize)
+                            {
+                                return "E";
+                            }
+
+                            remainLength = maxOutputSize - parts[0].Length - 1;
+                            // trim the fractional part gracefully. =
+                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
+                        }
+                        else
+                        {
+                            throw new DivideByZeroException("Divided by 0 ");
+                        }
+
+                }
+            }
+            catch (FormatException e)
+            {
+                Console.WriteLine("formatting failed: " + e);
+            }
+            catch (OverflowException e)
+            {
+                Console.WriteLine("Too many string to format to float: " + e);
+            }
+            catch (DivideByZeroException e)
+            {
+                Console.WriteLine("Attempt divided by 0: " + e);
+            }
+            return "E";
+        }
+
+        public string calculate(string oper, string firstOperand, string secondOperand, int maxOutputSize = 8)
+        {
+            try
+            {
+                switch (oper)
+                {
+                    case "+":
+                        return (Convert.ToDouble(firstOperand) + Convert.ToDouble(secondOperand)).ToString();
+                    case "-":
+                        return (Convert.ToDouble(firstOperand) - Convert.ToDouble(secondOperand)).ToString();
+                    case "X":
+                        return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand)).ToString();
+                    case "÷":
+                        // Not allow devide be zero
+                        if (secondOperand != "0")
+                        {
+                            double result;
+                            string[] parts;
+                            int remainLength;
+
+                            result = (Convert.ToDouble(firstOperand) / Convert.ToDouble(secondOperand));
+                            // split between integer part and fractional part
+                            parts = result.ToString().Split('.');
+                            // if integer part length is already break max output, return error
+                            if (parts[0].Length > maxOutputSize)
+                            {
+                                return "E";
+                            }
+
+                            remainLength = maxOutputSize - parts[0].Length - 1;
+                            // trim the fractional part gracefully. =
+                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
+                        }
+                        else
+                        {
+                            throw new DivideByZeroException("Can't divided by 0");
+                        }
+                    case "%":
+                        return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand) / 100).ToString();
+                }
+            }
+            catch (FormatException e)
+            {
+                Console.WriteLine("formatting failed: " + e);
+            }
+            catch (DivideByZeroException e)
+            {
+                Console.WriteLine("Divided by 0 " + e);
+            }
+            return "E";
+        }
+    }
+}
