+++ CPE200Lab1/CPE200Lab1/CalculatorEngine.cs
@@ -7,40 +7,24 @@ using System.Threading.Tasks;
 namespace CPE200Lab1
 {
 
-    public class CalculatorEngine
+    public class CalculatorEngine : BasicCalculatorEngine
     {
-        //change to public for using in RPNCalculatorengine
-        public bool isNumber(string str)
+        private Double firstOperand;
+        private Double secondOperand;
+
+        public void setFirstOperand(string num)
         {
-            double retNum;
-            return Double.TryParse(str, out retNum);
-        }
-        public bool isOperator(string str)
-        {
-            switch(str) {
-                case "+":
-                case "-":
-                case "X":
-                case "÷":
-                case "%":
-                    return true;
-            }
-            return false;
-        }
-        public bool isUnary(string str)
-        {
-            switch (str)
-            {
-                case "√":
-                case "1/x":
-                    return true;
-            }
-            return false;
+            firstOperand = Convert.ToDouble(num);
         }
 
-        public string Process(string str)
+        public void setSeccondOperand(string num)
         {
-            string[] parts = str.Split(' ');
+            secondOperand = Convert.ToDouble(num);
+        }
+
+        public string calculate(string oper)
+        {
+            string[] parts = oper.Split(' ');
             if(!(isNumber(parts[0]) && isOperator(parts[1]) && isNumber(parts[2])))
             {
                 return "E";
@@ -50,119 +34,6 @@ namespace CPE200Lab1
             }
 
         }
-        public string calculate(string operate, string operand, int maxOutputSize = 8)
-        {
-            try
-            {
-                switch (operate)
-                {
-                    case "√":
-                        {
-                            double result;
-                            string[] parts;
-                            int remainLength;
-
-                            result = Math.Sqrt(Convert.ToDouble(operand));
-                            // split between integer part and fractional part
-                            parts = result.ToString().Split('.');
-                            // if integer part length is already break max output, return error
-                            if (parts[0].Length > maxOutputSize)
-                            {
-                                return "E";
-                            }
-                            // calculate remaining space for fractional part.
-                            remainLength = maxOutputSize - parts[0].Length - 1;
-                            // trim the fractional part gracefully. 
-                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
-
-                        }
-                    case "1/x":
-                        if (operand != "0")
-                        {
-                            double result;
-                            string[] parts;
-                            int remainLength;
-
-                            result = (1.0 / Convert.ToDouble(operand));
-                            // split between integer part and fractional part
-                            parts = result.ToString().Split('.');
-                            // if integer part length is already break max output, return error
-                            if (parts[0].Length > maxOutputSize)
-                            {
-                                return "E";
-                            }
-
-                            remainLength = maxOutputSize - parts[0].Length - 1;
-                            // trim the fractional part gracefully. =
-                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
-                        }
-                        else
-                        {
-                            throw new DivideByZeroException("Divided by 0 ");
-                        }
-                        
-                }
-            }catch(FormatException e)
-            {
-                Console.WriteLine("formatting failed: " + e);
-            }catch(OverflowException e)
-            {
-                Console.WriteLine("Too many string to format to float: " + e);
-            }catch(DivideByZeroException e)
-            {
-                Console.WriteLine("Attempt divided by 0: " + e);
-            }
-            return "E";
-        }
-
-        public string calculate(string operate, string firstOperand, string secondOperand, int maxOutputSize = 8)
-        {
-            try
-            {
-                switch (operate)
-                {
-                    case "+":
-                        return (Convert.ToDouble(firstOperand) + Convert.ToDouble(secondOperand)).ToString();
-                    case "-":
-                        return (Convert.ToDouble(firstOperand) - Convert.ToDouble(secondOperand)).ToString();
-                    case "X":
-                        return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand)).ToString();
-                    case "÷":
-                        // Not allow devide be zero
-                        if (secondOperand != "0")
-                        {
-                            double result;
-                            string[] parts;
-                            int remainLength;
-
-                            result = (Convert.ToDouble(firstOperand) / Convert.ToDouble(secondOperand));
-                            // split between integer part and fractional part
-                            parts = result.ToString().Split('.');
-                            // if integer part length is already break max output, return error
-                            if (parts[0].Length > maxOutputSize)
-                            {
-                                return "E";
-                            }
-
-                            remainLength = maxOutputSize - parts[0].Length - 1;
-                            // trim the fractional part gracefully. =
-                            return result.ToString("N" + remainLength).TrimEnd('0', '.');
-                        }
-                        else
-                        {
-                            throw new DivideByZeroException("Can't divided by 0");
-                        }
-                    case "%":
-                        return (Convert.ToDouble(firstOperand) * Convert.ToDouble(secondOperand) / 100).ToString();
-                }
-            }catch(FormatException e)
-            {
-                Console.WriteLine("formatting failed: " + e);
-            }catch(DivideByZeroException e)
-            {
-                Console.WriteLine("Divided by 0 " + e);
-            }
-            return "E";
-        }
+   
     }
 }
